move types/rfc6962.go to types/types/rfc6962.go
# Conflicts:
#	internal/scti/handlers_test.go
