+++ internal/scti/chain_validation_test.go
@@ -24,7 +24,7 @@ import (
 	"time"
 
 	"github.com/transparency-dev/static-ct/internal/testdata"
-	"github.com/transparency-dev/static-ct/internal/types"
+	"github.com/transparency-dev/static-ct/internal/types/types"
 	"github.com/transparency-dev/static-ct/internal/x509util"
 )
